# src/routes/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Record_string.any_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"Did": {
				"type": "string",
				"example": "did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL"
			},
...
			"DidResolutionMetadata": {
				"properties": {
					"contentType": {
						"type": "string"
					},
					"error": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"enum": [
									"invalidDid",
									"notFound",
									"representationNotSupported",
									"unsupportedDidMethod"
								]
							}
						]
					},
					"message": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"DidDocument": {
				"$ref": "#/components/schemas/Record_string.any_"
			},
			"DIDDocumentMetadata": {
				"description": "Represents metadata about the DID document resulting from a {@link Resolvable.resolve} operation.",
				"properties": {
					"created": {
						"type": "string"
					},
					"updated": {
						"type": "string"
					},
					"deactivated": {
						"type": "boolean"
					},
					"versionId": {
						"type": "string"
					},
					"nextUpdate": {
						"type": "string"
					},
					"nextVersionId": {
						"type": "string"
					},
					"equivalentId": {
						"type": "string"
					},
					"canonicalId": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"DidDocumentMetadata": {
				"$ref": "#/components/schemas/DIDDocumentMetadata"
			},
			"DidResolutionResult": {
				"properties": {
					"didResolutionMetadata": {
						"$ref": "#/components/schemas/DidResolutionMetadata"
					},
					"didDocument": {
						"allOf": [
							{
								"$ref": "#/components/schemas/DidDocument"
							}
						],
						"nullable": true
					},
					"didDocumentMetadata": {
						"$ref": "#/components/schemas/DidDocumentMetadata"
					}
				},
				"required": [
					"didResolutionMetadata",
					"didDocument",
					"didDocumentMetadata"
				],
				"type": "object",
				"additionalProperties": false
			},
			"KeyType": {
				"enum": [
					"ed25519",
					"bls12381g1g2",
					"bls12381g1",
					"bls12381g2",
					"x25519",
					"p256",
					"p384",
					"p521"
				],
				"type": "string"
			},
			"PrivateKey": {
				"properties": {
					"keyType": {
						"$ref": "#/components/schemas/KeyType"
					},
					"privateKey": {
						"type": "string"
					}
				},
				"required": [
					"keyType",
					"privateKey"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ImportDidOptions": {
				"properties": {
					"did": {
						"type": "string"
					},
					"didDocument": {
						"$ref": "#/components/schemas/DidDocument"
					},
					"privateKeys": {
						"items": {
							"$ref": "#/components/schemas/PrivateKey"
						},
						"type": "array"
					},
					"overwrite": {
						"type": "boolean"
					}
				},
				"required": [
					"did"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DidCreateOptions": {
				"properties": {
					"method": {
						"type": "string"
					},
					"did": {
						"type": "string"
					},
					"options": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"secret": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"didDocument": {
						"$ref": "#/components/schemas/DidDocument"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@digicatapult/veritable-cloudagent",
		"version": "0.4.0",
		"description": "Rest endpoint wrapper for using your agent over HTTP",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {
			"name": "Digital Catapult"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/dids/{did}": {
			"get": {
				"operationId": "GetDidRecordByDid",
				"responses": {
					"200": {
						"description": "DidResolutionResult",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"resolveResult": {
													"$ref": "#/components/schemas/DidResolutionResult"
												}
											},
											"required": [
												"resolveResult"
											],
											"type": "object"
										},
										{
											"properties": {
												"resolveResult": {},
												"didDocumentMetadata": {
													"$ref": "#/components/schemas/DIDDocumentMetadata"
												},
												"didResolutionMetadata": {
													"$ref": "#/components/schemas/DidResolutionMetadata"
												},
												"didDocument": {
													"$ref": "#/components/schemas/Record_string.any_"
												}
											},
											"required": [
												"didDocumentMetadata",
												"didResolutionMetadata",
												"didDocument"
											],
											"type": "object"
										}
									]
								},
								"examples": {
									"Example 1": {
										"value": {
											"didDocument": {
												"@context": [
													"https://w3id.org/did/v1",
													"https://w3id.org/security/suites/ed25519-2018/v1",
													"https://w3id.org/security/suites/x25519-2019/v1"
												],
												"id": "did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL",
												"verificationMethod": [
													{
														"id": "did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL#z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL",
														"type": "Ed25519VerificationKey2018",
														"controller": "did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL",
														"publicKeyBase58": "6fioC1zcDPyPEL19pXRS2E4iJ46zH7xP6uSgAaPdwDrx"
													}
												],
												"authentication": [
													"did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL#z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL"
												],
												"assertionMethod": [
													"did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL#z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL"
												],
												"capabilityInvocation": [
													"did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL#z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL"
												],
												"capabilityDelegation": [
													"did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL#z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL"
												],
												"keyAgreement": [
													{
														"id": "did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL#z6LSrdqo4M24WRDJj1h2hXxgtDTyzjjKCiyapYVgrhwZAySn",
														"type": "X25519KeyAgreementKey2019",
														"controller": "did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL",
														"publicKeyBase58": "FxfdY3DCQxVZddKGAtSjZdFW9bCCW7oRwZn1NFJ2Tbg2"
													}
												]
											},
											"didDocumentMetadata": {},
											"didResolutionMetadata": {
												"contentType": "application/did+ld+json"
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Resolves did and returns did resolution result",
				"tags": [
					"Dids"
				],
				"security": [],
				"parameters": [
					{
						"description": "Decentralized Identifier",
						"in": "path",
						"name": "did",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Did"
						}
					}
				]
			}
		},
		"/dids/import": {
			"post": {
				"operationId": "ImportDid",
				"responses": {
					"200": {
						"description": "DidResolutionResultProps",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"didDocument": {
												"@context": [
													"https://w3id.org/did/v1",
													"https://w3id.org/security/suites/ed25519-2018/v1",
													"https://w3id.org/security/suites/x25519-2019/v1"
												],
												"id": "did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL",
												"verificationMethod": [
													{
														"id": "did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL#z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL",
														"type": "Ed25519VerificationKey2018",
														"controller": "did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL",
														"publicKeyBase58": "6fioC1zcDPyPEL19pXRS2E4iJ46zH7xP6uSgAaPdwDrx"
													}
												],
												"authentication": [
													"did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL#z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL"
												],
												"assertionMethod": [
													"did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL#z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL"
												],
												"capabilityInvocation": [
													"did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL#z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL"
												],
												"capabilityDelegation": [
													"did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL#z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL"
												],
												"keyAgreement": [
													{
														"id": "did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL#z6LSrdqo4M24WRDJj1h2hXxgtDTyzjjKCiyapYVgrhwZAySn",
														"type": "X25519KeyAgreementKey2019",
														"controller": "did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL",
														"publicKeyBase58": "FxfdY3DCQxVZddKGAtSjZdFW9bCCW7oRwZn1NFJ2Tbg2"
													}
												]
											},
											"didDocumentMetadata": {},
											"didResolutionMetadata": {
												"contentType": "application/did+ld+json"
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Import a Did to the Agent and return the did resolution result",
				"tags": [
					"Dids"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ImportDidOptions"
							}
						}
					}
				}
			}
		},
		"/dids/create": {
			"post": {
				"operationId": "CreateDid",
				"responses": {
					"200": {
						"description": "DidResolutionResultProps",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"did": "did:key:z6MkpGuzuD38tpgZKPfmLmmD8R6gihP9KJhuopMuVvfGzLmc",
											"didDocument": {
												"@context": [
													"https://w3id.org/did/v1",
													"https://w3id.org/security/suites/ed25519-2018/v1",
													"https://w3id.org/security/suites/x25519-2019/v1"
												],
												"id": "did:key:z6MkpGuzuD38tpgZKPfmLmmD8R6gihP9KJhuopMuVvfGzLmc",
												"verificationMethod": [
													{
														"id": "did:key:z6MkpGuzuD38tpgZKPfmLmmD8R6gihP9KJhuopMuVvfGzLmc#z6MkpGuzuD38tpgZKPfmLmmD8R6gihP9KJhuopMuVvfGzLmc",
														"type": "Ed25519VerificationKey2018",
														"controller": "did:key:z6MkpGuzuD38tpgZKPfmLmmD8R6gihP9KJhuopMuVvfGzLmc",
														"publicKeyBase58": "ApexJxnhZHC6Ctq4fCoNHKYgu87HuRTZ7oSyfehG57zE"
													}
												],
												"authentication": [
													"did:key:z6MkpGuzuD38tpgZKPfmLmmD8R6gihP9KJhuopMuVvfGzLmc#z6MkpGuzuD38tpgZKPfmLmmD8R6gihP9KJhuopMuVvfGzLmc"
												],
												"assertionMethod": [
													"did:key:z6MkpGuzuD38tpgZKPfmLmmD8R6gihP9KJhuopMuVvfGzLmc#z6MkpGuzuD38tpgZKPfmLmmD8R6gihP9KJhuopMuVvfGzLmc"
												],
												"keyAgreement": [
													{
														"id": "did:key:z6MkpGuzuD38tpgZKPfmLmmD8R6gihP9KJhuopMuVvfGzLmc#z6LSm5B4fB9NA55xB7PSeMYTMS9sf8uboJvyZBaDLLSZ7Ryd",
														"type": "X25519KeyAgreementKey2019",
														"controller": "did:key:z6MkpGuzuD38tpgZKPfmLmmD8R6gihP9KJhuopMuVvfGzLmc",
														"publicKeyBase58": "APzu8sLW4cND5j1g7i2W2qwPozNV6hkpgCrXqso2Q4Cs"
													}
												],
												"capabilityInvocation": [
													"did:key:z6MkpGuzuD38tpgZKPfmLmmD8R6gihP9KJhuopMuVvfGzLmc#z6MkpGuzuD38tpgZKPfmLmmD8R6gihP9KJhuopMuVvfGzLmc"
												],
												"capabilityDelegation": [
													"did:key:z6MkpGuzuD38tpgZKPfmLmmD8R6gihP9KJhuopMuVvfGzLmc#z6MkpGuzuD38tpgZKPfmLmmD8R6gihP9KJhuopMuVvfGzLmc"
												]
											},
											"secret": {}
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Create a Did and return the did resolution result",
				"tags": [
					"Dids"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DidCreateOptions"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}